fix(stringcalculator, outputview) : int 범위 초과 문제 해결을 위해 BigInteger 로 수정

+++ src/main/kotlin/calculator/view/OutputView.kt
@@ -1,8 +1,10 @@
 package calculator.view
 
+import java.math.BigInteger
+
 class OutputView {
 
-    fun printOutput(result : Int) {
+    fun printOutput(result : BigInteger) {
         println("결과 : $result")
     }
 }
